requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Vector Database & Indexing
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
hnswlib>=0.7.0

# Knowledge Graph & NLP
spacy>=3.5.0
networkx>=3.1
rdflib>=6.3.2

# Data Processing
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Evaluation Metrics
bert-score>=0.3.13
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
pickle5>=0.0.12

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Optional: For web demo
gradio>=3.35.0
fastapi>=0.100.0
uvicorn>=0.23.0